Reworking some verification tests into examples
